Fix float32 overflow in lambda_edd (BH mass × M_SUN overflows)
illustris_python returns float32 arrays. For BH_Mass > ~1e6 Msun,
float32(M_bh) × M_SUN (1.99e33) = ~1e41 overflows float32 max (3.4e38)
→ inf → lambda_edd = 0. Same root cause as eps_ff density overflow.

Fix: cast to float64 in eddington_mdot_msun_yr (safety net for all callers)
     and in load_halo BH arrays at load time (prevents all downstream issues).

+++ loaders.py
@@ -128,8 +128,8 @@ def load_halo(path, snapnum, halo_id, catalog, h, a, box_ckpch, bh_avail=None):
     n_bh = len(bh_raw["BH_Mass"]) if "BH_Mass" in bh_raw else 0
     bh = dict(
         pos    = U.comoving_to_physical(bh_raw["Coordinates"], a, h),
-        mass   = U.code_mass_to_msun(bh_raw["BH_Mass"],        h),
-        mdot   = U.bh_mdot_to_msun_yr(bh_raw["BH_Mdot"]),
+        mass   = U.code_mass_to_msun(bh_raw["BH_Mass"].astype(np.float64),   h),
+        mdot   = U.bh_mdot_to_msun_yr(bh_raw["BH_Mdot"].astype(np.float64)),
         egy_qm = bh_raw.get("BH_CumEgyInjection_QM", np.zeros(n_bh)),
         egy_rm = bh_raw.get("BH_CumEgyInjection_RM", np.zeros(n_bh)),
     )
